+++ src/Domain/Exception/AmbiguousGrade.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Climb\Grades\Domain\Exception;
-
-use DomainException;
-
-final class AmbiguousGrade extends DomainException {}